Add register tests after creating a program to parse the datasheet
The tables were programmatically extracted from the pdf and made into a
CSV and then this program was written to produce the mapping of
registers to their respective addresses. Because it is impractical to
perfectly map some of the variable names, there's a lookup table.

Also, correct a couple of addresses that were identified as incorrect.